% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenomStanModel-package.R
\docType{package}
\name{PhenomStanModel-package}
\alias{PhenomStanModel-package}
\alias{PhenomStanModel}
\title{The 'PhenomStanModel' package.}
\description{
A minimal package containing just one object - a compiled Stan model to be used in the Phenom Shiny Web App.
}
\references{
Stan Development Team (2021). RStan: the R interface to Stan. R package version 2.21.3. https://mc-stan.org
}
